modernize C-Fortran interface for Code/Flag table routines
